c map finished

+++ lab1.c
@@ -7,7 +7,7 @@ int main(int argc, char *argv[]){
         tot_len += strlen(argv[i]);
         printf("%-17s%4lu\n",argv[i],strlen(argv[i]));
     }
-    float avrg = (float)tot_len/argc;
+    float avrg = (float)tot_len/(float)argc;
     printf("\n%-17s%4u\n", "Total length", tot_len);
     printf("%-17s%4.2f\n", "Average length", avrg);
